feat: check if user is authorized to interact with conversation

+++ src/pages/api/conversations.ts
@@ -5,9 +5,18 @@ import type { NextApiRequest, NextApiResponse } from "next";
 
 const prisma = new PrismaClient();
 
+const isAuthorized = async (userId: string, conversationId: string) => {
+  const conversation = await prisma.conversation.findUnique({
+    where: { conversationId: conversationId },
+  });
+  if (!conversation) return false;
+  if (conversation.userId !== userId) return false; // Q: different http status code and error message?
+  return true;
+}
+
 const handler = async (req: NextApiRequest, rsp: NextApiResponse) => {
   const { userId } = getAuth(req);
-  if (!userId) return rsp.status(401).json({ error: "Unauthorized" });
+  if (!userId) return rsp.status(401).json({ error: "Unauthenticated" });
 
   if (req.method === "GET") {
     // get all conversations for user
@@ -21,13 +30,17 @@ const handler = async (req: NextApiRequest, rsp: NextApiResponse) => {
     const conversation = await prisma.conversation.create({ data: { userId: userId } });
     return rsp.status(200).json({ conversation });
   } else if (req.method === "PUT") {
-    // update conversation title
+    // get the conversation ID and new title from user
     const { conversationId } = req.query;
     if (!conversationId) return rsp.status(400).json({ error: "Conversation API handler: Missing conversationId" });
-
     const { newTitle } = req.body as { newTitle: string };
     if (!newTitle) return rsp.status(400).json({ error: "Conversation API handler: Missing newTitle" });
 
+    // check if the user is authorized to update the conversation
+    const authorized = await isAuthorized(userId, conversationId as string);
+    if (!authorized) return rsp.status(401).json({ error: "Unauthorized" });
+
+    // update the conversation title
     const conversation = await prisma.conversation.update({
       where: { conversationId: conversationId as string },
       data: { title: newTitle },
@@ -37,6 +50,10 @@ const handler = async (req: NextApiRequest, rsp: NextApiResponse) => {
     // delete conversation from database
     const { conversationId } = req.query;
     if (!conversationId) return rsp.status(400).json({ error: "Conversation API handler: Missing conversationId" });
+
+    // check if the user is authorized to delete the conversation
+    const authorized = await isAuthorized(userId, conversationId as string);
+    if (!authorized) return rsp.status(401).json({ error: "Unauthorized" });
 
     const conversation = await prisma.conversation.delete({
       where: { conversationId: conversationId as string }
